pyproject.toml
[project]
name = "sanity-agent"
version = "0.1.0"
description = "Sanity AI Therapist Agent - EuphorIA powered by LangChain + Gemini"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "pydantic>=2.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "sqlalchemy>=2.0.0",
    "pyodbc>=5.1.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]
